Move package of RDLUpdater and RQLResultSet
